aux: moved rest Serialization and ExceptionHandling and removed some models
