+++ src/main/java/sk/leziemevpezinku/spring/rest/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package sk.leziemevpezinku.spring.rest.mapper;
+package sk.leziemevpezinku.spring.rest.exception;
 
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.HttpStatus;
