+++ src/main/java/sk/leziemevpezinku/spring/rest/serialization/LocalDateDeserializer.java
@@ -1,4 +1,4 @@
-package sk.leziemevpezinku.spring.rest.mapper;
+package sk.leziemevpezinku.spring.rest.serialization;
 
 import com.fasterxml.jackson.core.JacksonException;
 import com.fasterxml.jackson.core.JsonParser;
